data-loader/requirements.txt
hazelcast-python-client[stats]~=5.2.0
